VietnameseASR/requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==24.1.0
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.8.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.1.0
audioread==3.0.1
bcrypt==4.2.1
bitarray==3.0.0
bitsandbytes==0.45.1
blosc2==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.12
datasets==2.14.6
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1737269734548/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
dill==0.3.9
distlib==0.3.9
distro==1.7.0
docker-pycreds==0.4.0
editdistance==0.8.1
einops==0.6.1
evaluate==0.4.1
examples==1.0.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
-e git+https://github.com/facebookresearch/fairseq@4c6b689eebe66a53717dacf28cba7a11b6ffa64f#egg=fairseq
fastBPE==0.1.0
ffmpeg-python==0.2.0
filelock==3.17.0
fire==0.5.0
flashlight-sequence @ git+https://github.com/flashlight/sequence.git@158f4670379a6885bb5d7f7ee8dabd4fb548d4c5
flashlight-text==0.0.7
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.4.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.4
httpx==0.24.0
huggingface-hub==0.28.1
hydra-core==1.0.7
HyperPyYAML==1.1.0
idna==3.10
idx2numpy==1.2.3
importlib-resources==6.1.0
importlib_metadata==8.4.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_ipython_1738421264/work
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.5
jiwer==3.0.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kaldi-io==0.9.8
kenlm @ git+https://github.com/kpu/kenlm.git@f6c947dc943859e265fabce886232205d0fb2b37
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
Levenshtein==0.26.1
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.1.0
lightning-utilities==0.9.0
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mindspore==2.5.0
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
ndindex==1.9.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nltk==3.8.1
numba==0.61.0
numexpr==2.10.2
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.1.2
paramiko==3.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
peft==0.6.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1733301927746/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==5.28.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1735327341346/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1733302279685/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl#sha256=92c32ff62b5fd8cf325bec5ab90d7be3d2a8ca8c8a3813ff487a8d2002630d1f
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==14.0.0
pybind11==2.13.6
pycocoevalcap @ git+https://github.com/salaniz/pycocoevalcap.git@a24f74c408c918f1f4ec34e9514bc8a76ce41ffd
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyDeprecate==0.3.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.2.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.10
pytz==2023.3.post1
pyvi==0.1.1
PyYAML==6.0.2
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1738270958168/work
RapidFuzz==3.12.1
rarfile==4.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacrebleu==2.5.1
safetensors==0.5.2
sanic==0.7.0
scikit-learn==1.2.2
scipy==1.15.1
SecretStorage==3.3.1
sentencepiece==0.1.99
sentry-sdk==2.21.0
setproctitle==1.3.4
six==1.17.0
sklearn-crfsuite==0.5.0
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==0.5.15
srt==3.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.1
tables==3.10.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
TensorFlowASR @ file:///data/npl/Speech2Text/TensorFlowASR-main
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==0.6.13
tokenizers==0.15.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.6.1
torchvision==0.20.1+cu121
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1732615898999/work
tqdm==4.66.6
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.37.2
triton==3.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
ujson==5.10.0
urllib3==2.3.0
uvloop==0.21.0
virtualenv==20.29.2
visdom==0.2.4
vosk==0.3.45
wadllib==1.3.6
wandb==0.19.6
warp-rnnt==0.7.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.0
wget==3.2
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
